+++ InfinityTools/commands/ToolExport.cs
@@ -1,3 +1,4 @@
+using System.Linq;
 using InfinityHammer;
 using ServerDevcommands;
 using UnityEngine;
@@ -8,16 +9,21 @@ public class ToolExportCommand
 {
   public ToolExportCommand()
   {
-    Helper.Command("tool_export", "[hammer/hoe/etc] [name] - Export a tool.", Execute);
-    AutoComplete.Register("tool_export", (int index, int subIndex) => null);
+    Helper.Command("tool_export", "[name] - Exports a tool", Execute);
+    AutoComplete.Register("tool_export", (int index, int subIndex) => ToolManager.GetAll().Select(tool => tool.Name).ToList());
   }
   protected static void Execute(Terminal.ConsoleEventArgs args)
   {
-    Helper.ArgsCheck(args, 2, "Missing the equipment.");
-    Helper.ArgsCheck(args, 3, "Missing the tool.");
-    var equipment = args[1];
-    var name = string.Join(" ", args.Args, 2, args.Length - 1);
-    GUIUtility.systemCopyBuffer = ToolManager.Export(equipment, name);
-    HammerHelper.Message(args.Context, $"Export tool {name} to the clipboard.");
+    Helper.ArgsCheck(args, 2, "Missing the tool name.");
+    var tool = string.Join(" ", args.Args, 1, args.Length - 1);
+    var equipment = Hammer.Get();
+    var result = ToolManager.Export(equipment, tool);
+    if (string.IsNullOrEmpty(result))
+    {
+      HammerHelper.Message(args.Context, $"Unable to find the tool {tool}.");
+      return;
+    }
+    GUIUtility.systemCopyBuffer = result;
+    HammerHelper.Message(args.Context, $"Exported tool {tool} to the clipboard.");
   }
 }